HSEARCH-1316 Update Spring ITs to use Spring Boot
