+++ integrationtest/spring/src/test/java/org/hibernate/search/test/integration/jtaspring/JtaAndSpringApplicationConfiguration.java
@@ -1,0 +1,25 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.test.integration.jtaspring;
+
+import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * @author Yoann Rodiere
+ */
+@Configuration
+@SpringBootConfiguration
+@EnableAutoConfiguration
+@ComponentScan
+@EntityScan
+public class JtaAndSpringApplicationConfiguration {
+
+}
