+++ integrationtest/spring/src/test/java/org/hibernate/search/test/integration/jtaspring/dao/BoxDAO.java
@@ -4,11 +4,12 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test.integration.jtaspring;
+package org.hibernate.search.test.integration.jtaspring.dao;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.hibernate.search.test.integration.jtaspring.entity.Box;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
