+++ integrationtest/spring/src/test/java/org/hibernate/search/test/integration/jtaspring/dao/SnertDAO.java
@@ -4,31 +4,33 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test.integration.jtaspring;
+package org.hibernate.search.test.integration.jtaspring.dao;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.hibernate.search.test.integration.jtaspring.entity.Snert;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Repository
-public class MuffinDAO {
+public class SnertDAO {
+
 	@PersistenceContext
 	private EntityManager entityManager;
 
 	@Transactional
-	public void persist(Muffin muffin) {
-		entityManager.persist( muffin );
+	public void persist(Snert snert) {
+		entityManager.persist( snert );
 	}
 
 	@Transactional
-	public void merge(Muffin muffin) {
-		entityManager.merge( muffin );
+	public void merge(Snert snert) {
+		entityManager.merge( snert );
 	}
 
 	@Transactional
-	public void remove(Muffin muffin) {
-		entityManager.remove( muffin );
+	public void remove(Snert snert) {
+		entityManager.remove( snert );
 	}
 }
