--- integrationtest/spring/src/test/resources/persistence.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
- ~ Hibernate Search, full-text search for your domain model
- ~
- ~ License: GNU Lesser General Public License (LGPL), version 2.1 or later
- ~ See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
-  -->
-<persistence xmlns="http://java.sun.com/xml/ns/persistence" version="1.0">
-
-    <persistence-unit name="integration" transaction-type="JTA">
-        <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
-        <class>org.hibernate.search.test.integration.jtaspring.Snert</class>
-        <class>org.hibernate.search.test.integration.jtaspring.Container</class>
-        <class>org.hibernate.search.test.integration.jtaspring.Box</class>
-        <class>org.hibernate.search.test.integration.jtaspring.Doughnut</class>
-        <class>org.hibernate.search.test.integration.jtaspring.Muffin</class>
-    </persistence-unit>
-
-    <persistence-unit name="jbossjta" transaction-type="JTA">
-        <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
-        <properties>
-            <property name="" value="" />
-        </properties>
-    </persistence-unit>
-
-</persistence>
